build_cmake.sh - Fix ufw bash error prep_host.sh- Add 'install' arg to tools build script

+++ build_cmake.sh
@@ -67,7 +67,9 @@ if [ "$sync" = "true" ]; then
     fi
 fi
 
-rm -r $cmake_build
+if [ -f $cmake_build ]; then
+    rm -r $cmake_build
+fi
 mkdir -p $cmake_build
 checkfail $? "Failed: mkdir $cmake_build"
